# openapi/raibitserver.yaml
openapi: 3.1.0
info:
  title: RAIBITSERVER API
  version: 0.1.0
  description: Desired-state API contract for the RAIBITSERVER PaaS + DBaaS control plane.
servers:
  - url: https://api.raibitserver.app
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/services/{serviceId}/deployments:
    post:
      summary: Queue a deployment workflow for a service in a scoped project.
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: serviceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '202': { description: Deployment and workflow job queued }
  /projects/{projectId}/services/{serviceId}/env:
    get:
      summary: List masked environment variables for a service.
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: serviceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Masked service environment }
    post:
      summary: Upsert service environment variables and secret values.
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: serviceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Environment variables stored with secrets masked }
  /projects/{projectId}/services/{serviceId}/env-file:
    post:
      summary: Upload .env file content for a service.
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: serviceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Parsed .env entries stored with secret masking }
  /integrations/github:
    get:
      summary: List GitHub integrations visible to an organization-scoped subject.
      parameters:
        - name: organizationId
          in: query
          required: false
          schema: { type: string }
      responses:
        '200': { description: GitHub integrations }
    post:
      summary: Connect a GitHub token/App installation to an organization.
      responses:
        '201': { description: GitHub integration stored with token masked }
  /projects/{projectId}/services/{serviceId}/github:
    post:
      summary: Attach a GitHub repository/integration to a service desired state.
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: string }
        - name: serviceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Service source updated to GitHub repository }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...